ECMS-6518 implement REMOTE_09 for File Explorer
