+++ core/webui-explorer/src/main/java/org/exoplatform/ecm/webui/component/explorer/rightclick/manager/OpenDocumentManageComponent.java
@@ -73,7 +73,7 @@ public class OpenDocumentManageComponent extends UIAbstractManagerComponent {
               +httpServletRequest.getServerPort() + "/"
               + WCMCoreUtils.getRestContextName()+ "/private/jcr/" + repo + "/" + ws + nodePath;
 
-      if(!PermissionUtil.canSetProperty(currentNode) && currentNode.isLocked()){
+      if(!PermissionUtil.canSetProperty(currentNode) || currentNode.isLocked()){
         String[] userLock = {currentNode.getLock().getLockOwner()};
 
         UIWorkingArea uiWorkingArea = event.getSource().getParent();
