+++ core/webui/src/main/java/org/exoplatform/ecm/webui/utils/PermissionUtil.java
@@ -16,93 +16,10 @@
  */
 package org.exoplatform.ecm.webui.utils;
 
-import java.security.AccessControlException;
-
-import javax.jcr.Node;
-import javax.jcr.RepositoryException;
-
-import org.exoplatform.services.jcr.access.PermissionType;
-import org.exoplatform.services.jcr.core.ExtendedNode;
-import org.exoplatform.services.security.IdentityConstants;
 
 /**
  * The Class PermissionUtil use to check permission for a node
  */
-public class PermissionUtil {
-
-  /**
-   * Can read node
-   *
-   * @param node the node
-   * @return true, if successful
-   * @throws RepositoryException the repository exception
-   */
-  public static boolean canRead(Node node) throws RepositoryException {
-    return checkPermission(node,PermissionType.READ);
-  }
-
-  /**
-   * Can add node.
-   *
-   * @param node the node
-   * @return true, if successful
-   * @throws RepositoryException the repository exception
-   */
-  public static boolean canAddNode(Node node) throws RepositoryException {
-    return checkPermission(node,PermissionType.ADD_NODE);
-  }
-
-  /**
-   * Can change permission.
-   *
-   * @param node the node
-   * @return true, if successful
-   * @throws RepositoryException the repository exception
-   */
-  public static boolean canChangePermission(Node node) throws RepositoryException {
-    return checkPermission(node,PermissionType.CHANGE_PERMISSION);
-  }
-
-  /**
-   * Checks if is any role.
-   *
-   * @param node the node
-   * @return true, if is any role
-   * @throws RepositoryException the repository exception
-   */
-  public static boolean isAnyRole(Node node)throws RepositoryException {
-    return checkPermission(node,IdentityConstants.ANY);
-  }
-
-  /**
-   * Can set property.
-   *
-   * @param node the node
-   * @return true, if successful
-   * @throws RepositoryException the repository exception
-   */
-  public static boolean canSetProperty(Node node) throws RepositoryException {
-    return checkPermission(node,PermissionType.SET_PROPERTY);
-  }
-
-  /**
-   * Can remove node.
-   *
-   * @param node the node
-   * @return true, if successful
-   * @throws RepositoryException the repository exception
-   */
-  public static boolean canRemoveNode(Node node) throws RepositoryException {
-    return checkPermission(node,PermissionType.REMOVE);
-  }
-
-  private static boolean checkPermission(Node node,String permissionType) throws RepositoryException {
-    try {
-      ((ExtendedNode)node).checkPermission(permissionType);
-      return true;
-    } catch(AccessControlException e) {
-      return false;
-    }
-  }
+public class PermissionUtil extends org.exoplatform.ecm.utils.permission.PermissionUtil {
 
 }
